--- art/defences/postprocess/__init__.py
@@ -1,9 +1,0 @@
-"""
-Module implementing postprocessing defences against adversarial attacks.
-"""
-from art.defences.postprocess.class_labels import ClassLabels
-from art.defences.postprocess.gaussian_noise import GaussianNoise
-from art.defences.postprocess.high_confidence import HighConfidence
-from art.defences.postprocess.postprocessor import Postprocessor
-from art.defences.postprocess.reverse_sigmoid import ReverseSigmoid
-from art.defences.postprocess.rounded import Rounded
