+++ art/defences/postprocessor/__init__.py
@@ -1,0 +1,9 @@
+"""
+Module implementing postprocessing defences against adversarial attacks.
+"""
+from art.defences.postprocessor.class_labels import ClassLabels
+from art.defences.postprocessor.gaussian_noise import GaussianNoise
+from art.defences.postprocessor.high_confidence import HighConfidence
+from art.defences.postprocessor.postprocessor import Postprocessor
+from art.defences.postprocessor.reverse_sigmoid import ReverseSigmoid
+from art.defences.postprocessor.rounded import Rounded
