+++ art/defences/postprocessor/gaussian_noise.py
@@ -16,37 +16,40 @@
 # TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 """
-This module implements confidence added to the classifier output.
+This module implements Gaussian noise added to the classifier output.
 """
 import logging
 
-from art.defences.postprocess.postprocessor import Postprocessor
+import numpy as np
+
+from art.defences.postprocessor.postprocessor import Postprocessor
+from art.utils import is_probability
 
 logger = logging.getLogger(__name__)
 
 
-class HighConfidence(Postprocessor):
+class GaussianNoise(Postprocessor):
     """
-    Implementation of a postprocessor based on selecting high confidence predictions to return as classifier output.
+    Implementation of a postprocessor based on adding Gaussian noise to classifier output.
     """
-    params = ['cutoff']
+    params = ['scale']
 
-    def __init__(self, cutoff=0.25, apply_fit=False, apply_predict=True):
+    def __init__(self, scale=0.2, apply_fit=False, apply_predict=True):
         """
-        Create a HighConfidence postprocessor.
+        Create a GaussianNoise postprocessor.
 
-        :param cutoff: Minimal value for returned prediction output.
-        :type cutoff: `float`
+        :param scale: Standard deviation of the distribution.
+        :type scale: `float`
         :param apply_fit: True if applied during fitting/training.
         :type apply_fit: `bool`
         :param apply_predict: True if applied during predicting.
         :type apply_predict: `bool`
         """
-        super(HighConfidence, self).__init__()
+        super(GaussianNoise, self).__init__()
         self._is_fitted = True
         self._apply_fit = apply_fit
         self._apply_predict = apply_predict
-        self.set_params(cutoff=cutoff)
+        self.set_params(scale=scale)
 
     @property
     def apply_fit(self):
@@ -65,8 +68,25 @@ class HighConfidence(Postprocessor):
         :return: Postprocessed model output.
         :rtype: `np.ndarray`
         """
+        # Generate random noise
+        noise = np.random.normal(loc=0.0, scale=self.scale, size=preds.shape)
+
+        # Add noise to model output
         post_preds = preds.copy()
-        post_preds[post_preds < self.cutoff] = 0.0
+        post_preds += noise
+
+        if preds.shape[1] > 1:
+            # Check if model output is logit or probability
+            are_probability = [is_probability(x) for x in preds]
+            all_probability = np.sum(are_probability) == preds.shape[0]
+
+            # Finally normalize probability output
+            if all_probability:
+                post_preds[post_preds < 0.0] = 0.0
+                sums = np.sum(post_preds, axis=1)
+                post_preds /= sums
+        else:
+            post_preds[post_preds < 0.0] = 0.0
 
         return post_preds
 
@@ -80,14 +100,14 @@ class HighConfidence(Postprocessor):
         """
         Take in a dictionary of parameters and apply checks before saving them as attributes.
 
-        :param cutoff: Minimal value for returned prediction output.
-        :type cutoff: `float`
+        :param scale: Standard deviation of the distribution.
+        :type scale: `float`
         :return: `True` when parsing was successful
         """
         # Save defence-specific parameters
-        super(HighConfidence, self).set_params(**kwargs)
+        super(GaussianNoise, self).set_params(**kwargs)
 
-        if self.cutoff <= 0:
-            raise ValueError('Minimal value must be positive.')
+        if self.scale <= 0:
+            raise ValueError('Standard deviation must be positive.')
 
         return True
