+++ art/defences/preprocessor/__init__.py
@@ -1,8 +1,6 @@
 """
-Module implementing multiple types of defences against adversarial attacks.
+Module implementing preprocessing defences against adversarial attacks.
 """
-from art.defences.trainer.adversarial_trainer import AdversarialTrainer
-
 from art.defences.preprocessor.feature_squeezing import FeatureSqueezing
 from art.defences.preprocessor.gaussian_augmentation import GaussianAugmentation
 from art.defences.preprocessor.jpeg_compression import JpegCompression
@@ -12,10 +10,3 @@ from art.defences.preprocessor.preprocessor import Preprocessor
 from art.defences.preprocessor.spatial_smoothing import SpatialSmoothing
 from art.defences.preprocessor.thermometer_encoding import ThermometerEncoding
 from art.defences.preprocessor.variance_minimization import TotalVarMin
-
-from art.defences.postprocessor.class_labels import ClassLabels
-from art.defences.postprocessor.gaussian_noise import GaussianNoise
-from art.defences.postprocessor.high_confidence import HighConfidence
-from art.defences.postprocessor.postprocessor import Postprocessor
-from art.defences.postprocessor.reverse_sigmoid import ReverseSigmoid
-from art.defences.postprocessor.rounded import Rounded
