+++ art/defences/preprocessor/jpeg_compression.py
@@ -31,7 +31,7 @@ import logging
 
 import numpy as np
 
-from art.defences.preprocess.preprocessor import Preprocessor
+from art.defences.preprocessor.preprocessor import Preprocessor
 from art.config import ART_NUMPY_DTYPE
 
 logger = logging.getLogger(__name__)
