+++ art/defences/preprocessor/spatial_smoothing.py
@@ -32,7 +32,7 @@ import logging
 import numpy as np
 from scipy import ndimage
 
-from art.defences.preprocess.preprocessor import Preprocessor
+from art.defences.preprocessor.preprocessor import Preprocessor
 from art.config import ART_NUMPY_DTYPE
 
 logger = logging.getLogger(__name__)
