+++ art/defences/preprocessor/thermometer_encoding.py
@@ -30,7 +30,7 @@ import logging
 
 import numpy as np
 
-from art.defences.preprocess.preprocessor import Preprocessor
+from art.defences.preprocessor.preprocessor import Preprocessor
 from art.utils import to_categorical
 from art.config import ART_NUMPY_DTYPE
