+++ art/defences/preprocessor/variance_minimization.py
@@ -31,7 +31,7 @@ import logging
 import numpy as np
 from scipy.optimize import minimize
 
-from art.defences.preprocess.preprocessor import Preprocessor
+from art.defences.preprocessor.preprocessor import Preprocessor
 from art.config import ART_NUMPY_DTYPE
 
 logger = logging.getLogger(__name__)
